move get_type_from_annotation to context.types
