+++ vyper/context/definitions/__init__.py
@@ -3,7 +3,6 @@ from vyper.context.definitions.functions import (  # NOQA: F401
 )
 from vyper.context.definitions.utils import (  # NOQA: F401
     get_definition_from_node,
-    get_index_value,
     get_literal_or_raise,
 )
 from vyper.context.definitions.values import (  # NOQA: F401
