+++ vyper/context/definitions/functions.py
@@ -20,7 +20,6 @@ from vyper.context.definitions.bases import (
 )
 from vyper.context.definitions.utils import (
     get_definition_from_node,
-    get_type_from_annotation,
 )
 from vyper.context.definitions.values import (
     Reference,
@@ -28,6 +27,7 @@ from vyper.context.definitions.values import (
 )
 from vyper.context.types import (
     get_builtin_type,
+    get_type_from_annotation,
 )
 from vyper.context.utils import (
     compare_types,
